backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
pandas==2.2.3
structlog==25.3.1
emergentintegrations==0.1.0